Fix markdownify filter to use main md library
