+++ lib/slugify.js
@@ -1,0 +1,3 @@
+const string = require('string')
+
+module.exports = s => string(s).replaceAll('.', '-').slugify().toString()
